+++ ksword/src/commonMain/kotlin/org/crosswire/ksword/passage/Key.kt
@@ -47,7 +47,7 @@ interface Key : Comparable<Key>, Iterable<Key> /*, kotlin.Cloneable, java.io.Ser
      * The key to use to cut down unnecessary output.
      * @return The string representation
      */
-    fun getName(base: Key?): String?
+    fun getName(base: Key?): String
 
     /**
      * A Human readable version of the Key's top level name. For Biblical
@@ -56,7 +56,7 @@ interface Key : Comparable<Key>, Iterable<Key> /*, kotlin.Cloneable, java.io.Ser
      *
      * @return a String containing a description of the Key
      */
-    fun getRootName(): String?
+    fun getRootName(): String
 
     /**
      * The OSIS defined reference specification for this Key. When the key is a
@@ -66,7 +66,7 @@ interface Key : Comparable<Key>, Iterable<Key> /*, kotlin.Cloneable, java.io.Ser
      *
      * @return a String containing the OSIS description of the verses
      */
-    fun getOsisRef(): String?
+    fun getOsisRef(): String
 
     /**
      * The OSIS defined id specification for this Key. When the key is a single
@@ -75,7 +75,7 @@ interface Key : Comparable<Key>, Iterable<Key> /*, kotlin.Cloneable, java.io.Ser
      *
      * @return a String containing the OSIS description of the verses
      */
-    fun getOsisID(): String?
+    fun getOsisID(): String
 
     /**
      * All keys have parents unless they are the root of a Key.